create index toko (still progress)

+++ database/migrations/2020_08_28_115541_toko.php
@@ -34,7 +34,7 @@ class Toko extends Migration
             $table->softDeletes();
             $table->timestamps();
 
-            $table->unsignedInteger('rating')->nullable();
+            $table->float('rating')->nullable();
             $table->enum('status', ['tidak aktif', 'aktif']);
             $table->enum('verifikasi', ['belum terverifikasi', 'terverifikasi']);
